chore(shared): rename i18n to translations
